Convert Karma configs to TS for easier type checking
